aply exception

+++ obs_scripts/ROS_cross_point.py
@@ -28,9 +28,9 @@ p.add_argument('--obsfile', type=str,
                help='absolute path for obsfile', required=True)
 p.add_argument('--tau', type=float,
                help='tau. default=%.1f'%(tau))
-p.add_argument('--integmin', type=float,
+p.add_argument('--integmin', type=int,
                help='integrange_min')
-p.add_argument('--integmax', type=float,
+p.add_argument('--integmax', type=int,
                help='integrange_max')
 
 args = p.parse_args()
